+++ components/src/Table/Table.story.js
@@ -94,9 +94,9 @@ const rowData = [
   { date: "2019-10-03", expectedQuantity: "2,475 eaches", actualQuantity: "1,425 eaches", id: "r3" },
   { date: "2019-10-04", expectedQuantity: "2,475 eaches", actualQuantity: "675 eaches", id: "r4" },
   { date: "2019-10-07", expectedQuantity: "2,475 eaches", actualQuantity: "1,575 eaches", id: "r5" },
-  { date: "2019-10-22", expectedQuantity: "1,725 eaches", actualQuantity: "-", id: "r6" },
-  { date: "2019-10-23", expectedQuantity: "2,475 eaches", actualQuantity: "-", id: "r7" },
-  { date: "2019-10-24", expectedQuantity: "2,475 eaches", actualQuantity: "-", id: "r8" }
+  { date: "2019-10-22", expectedQuantity: "1,725 eaches", actualQuantity: "-", id: "r7" },
+  { date: "2019-10-23", expectedQuantity: "2,475 eaches", actualQuantity: "-", id: "r8" },
+  { date: "2019-10-24", expectedQuantity: "2,475 eaches", actualQuantity: "-", id: "r9" }
 ];
 
 const rowDataWithWidths = [
@@ -158,8 +158,8 @@ const rowDataWithEverything = [
 ];
 
 const footerRowData = [
-  { date: "Total", expectedQuantity: "18,000 eaches", actualQuantity: "7,725 eaches", id: "r1" },
-  { date: "Attainment", expectedQuantity: "", actualQuantity: "41.5%", id: "r2" }
+  { date: "Total", expectedQuantity: "18,000 eaches", actualQuantity: "7,725 eaches" },
+  { date: "Attainment", expectedQuantity: "", actualQuantity: "41.5%" }
 ];
 
 storiesOf("Table", module)
@@ -214,9 +214,9 @@ storiesOf("Table", module)
   ))
   .add("with a footer", () => (
     <>
-      <Table columns={columns} rows={rowData} footerRows={footerRowData} />
+      <Table columns={columns} keyField="date" rows={rowData} footerRows={footerRowData} />
       <Text mt="x6">Loading state:</Text>
-      <Table columns={columns} rows={rowData} footerRows={footerRowData} loading />
+      <Table columns={columns} keyField="date" rows={rowData} footerRows={footerRowData} loading />
     </>
   ))
   .add("with lots of rows and columns", () => (